Added options to parser
Options:
    -include/exclude locations
    -switch on/off printing to console
Options are passed in object as second argument to parse function.

Because in Jison there is no support for passing arguments to parser
generated as a program, I made required hacking in Makehere script.

Now you can call it with arguments:
    -locations/nolocations
    -print/noprint

Everywhere the defaults are { locations: true, print: false }.
They are specified in the parser's grammar file.